Clickable frames now functional.
Clickable frames now detect hovers, hover end, and clicks.

Prints to console.

5 Cards succesfully on screen and react to interaction.

Next: Click and drag.

+++ kitten/mouse picking/MousePicker.h
@@ -8,7 +8,7 @@
 namespace MousePicker
 {
 	bool rayHits(const kitten::Ray& p_ray, const kitten::ClickableBox* p_clickable, double* p_hitOutput);	
-	bool uiHit(const kitten::ClickableFrame* p_clickable, int p_mouseX, int p_mouseY);
+	bool uiHit(const kitten::ClickableFrame* p_clickable, float p_mouseX, float p_mouseY);
 	kitten::ClickableBox* getClosestHit(const kitten::Ray& p_ray);
 	kitten::ClickableFrame* getClosestHitFrame(int mouseX, int mouseY);
 }
